# api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: Wesley's Ambacht Catering API
  description: |
    # Wesley's Ambacht Catering API

    A comprehensive REST API for managing catering bookings, quotes, availability, and add-on services for Wesley's Ambacht, a premium Dutch catering company.

    ## Features

    - **Real-time Availability Management** - Live booking slots with automatic reservation
    - **Dynamic Quote Generation** - Intelligent pricing based on service tiers and guest counts
    - **Add-on Services** - Flexible service extensions with per-person and flat-rate pricing
    - **Multi-tier Service Categories** - Essential, Premium, and Luxury tiers for different events
    - **Real-time Updates** - WebSocket subscriptions for live booking and availability updates
    - **Comprehensive Logging** - Full request/response logging with performance metrics

    ## Authentication

    This API uses Supabase Row Level Security (RLS) with JWT tokens. Include your authorization token in the `Authorization` header:

    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## Rate Limiting

    - 100 requests per minute for general endpoints
    - 500 requests per minute for availability checks
    - 50 requests per minute for booking creation

    ## Error Handling

    All endpoints return structured error responses:

    ```json
    {
      "error": {
        "code": "VALIDATION_ERROR",
        "message": "Guest count must be between 10 and 200",
        "details": {
...
        }
...

...
servers:
...
paths:
  /availability:
    get:
      summary: Get availability slots
      description: |
        Retrieve availability slots for a date range. Returns all slots with booking counts and availability status.

        **Performance**: Results are cached for 5 minutes. Use real-time subscriptions for live updates.
      operationId: getAvailabilitySlots
      tags:
        - Availability
      parameters:
        - name: start_date
          in: query
          description: Start date for availability range (YYYY-MM-DD)
          required: false
          schema:
            type: string
            format: date
            example: "2024-03-01"
        - name: end_date
          in: query
          description: End date for availability range (YYYY-MM-DD)
          required: false
          schema:
            type: string
            format: date
            example: "2024-03-31"
      responses:
        "200":
          description: Availability slots retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AvailabilitySlot"
                  meta:
                    $ref: "#/components/schemas/PaginationMeta"
              examples:
                success:
                  summary: Successful availability query
                  value:
                    data:
                      - id: "550e8400-e29b-41d4-a716-446655440000"
                        date: "2024-03-15"
                        time_slot: "18:00"
                        max_bookings: 3
                        current_bookings: 1
                        is_blocked: false
                        created_at: "2024-01-15T10:00:00Z"
                        updated_at: "2024-01-15T14:30:00Z"
                    meta:
                      total: 45
                      page: 1
                      per_page: 50
  /availability/check:
    post:
      summary: Check specific slot availability
      description: |
        Check if a specific date and time slot is available for booking.

        **Real-time**: This endpoint provides real-time availability status.
      operationId: checkAvailability
      tags:
        - Availability
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - date
                - time
              properties:
                date:
                  type: string
                  format: date
                  description: Date to check (YYYY-MM-DD)
                  example: "2024-03-15"
                time:
                  type: string
                  pattern: "^[0-2][0-9]:[0-5][0-9]$"
                  description: Time slot to check (HH:MM)
                  example: "18:00"
      responses:
        "200":
          description: Availability status
          content:
            application/json:
              schema:
                type: object
                properties:
                  available:
                    type: boolean
                  remaining_slots:
                    type: integer
                  time_slot:
                    type: string
              example:
                available: true
                remaining_slots: 2
                time_slot: "18:00"
  /availability/slots/{date}:
    get:
      summary: Get available time slots for a specific date
      description: |
        Get all available time slots for a specific date, excluding fully booked or blocked slots.
      operationId: getAvailableTimeSlots
      tags:
        - Availability
      parameters:
        - name: date
          in: path
          required: true
          description: Date to get time slots for (YYYY-MM-DD)
          schema:
            type: string
            format: date
            example: "2024-03-15"
      responses:
        "200":
          description: Available time slots
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AvailabilitySlot"
                  grouped_slots:
                    type: object
                    properties:
                      morning:
                        type: array
                        items:
                          type: string
                      afternoon:
                        type: array
                        items:
                          type: string
                      evening:
                        type: array
                        items:
                          type: string
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    AvailabilitySlot:
      type: object
      required:
        - id
        - date
        - time_slot
        - max_bookings
        - current_bookings
        - is_blocked
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        date:
          type: string
          format: date
          description: Date of the availability slot
        time_slot:
          type: string
          pattern: "^[0-2][0-9]:[0-5][0-9]$"
          description: Time slot in HH:MM format
        max_bookings:
          type: integer
          minimum: 1
          description: Maximum number of concurrent bookings allowed
        current_bookings:
          type: integer
          minimum: 0
          description: Current number of bookings for this slot
        is_blocked:
          type: boolean
          description: Whether this slot is manually blocked
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      example:
        id: "550e8400-e29b-41d4-a716-446655440000"
        date: "2024-03-15"
        time_slot: "18:00"
        max_bookings: 3
        current_bookings: 1
        is_blocked: false
        created_at: "2024-01-15T10:00:00Z"
        updated_at: "2024-01-15T14:30:00Z"
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          description: Total number of items
        page:
          type: integer
          description: Current page number
        per_page:
          type: integer
          description: Items per page
        total_pages:
          type: integer
          description: Total number of pages
        has_next:
          type: boolean
          description: Whether there are more pages
        has_prev:
          type: boolean
          description: Whether there are previous pages
security:
  - BearerAuth: []
...